refactor: 拆分平台模型与QQ适配
