+++ qq-maid-gateway-rs/src/respond.rs
@@ -14,7 +14,7 @@ use tracing::{debug, info, warn};
 
 use crate::{
     event::{C2cMessage, GroupMessage},
-    gateway::platform::InboundMessage,
+    gateway::platform,
     logging::mask_openid,
 };
 
@@ -179,44 +179,41 @@ fn log_core_output_success(
 }
 
 pub fn core_request_from_c2c_message(message: &C2cMessage, content: String) -> CoreRequest {
-    let inbound = InboundMessage::from_qq_c2c(message);
-    inbound
-        .to_core_request(content)
+    let inbound = platform::qq_official::inbound_from_c2c(message);
+    platform::to_core_request(&inbound, content)
         .expect("QQ C2C inbound message should map to CoreRequest")
 }
 
 pub fn core_request_from_group_message(message: &GroupMessage, content: String) -> CoreRequest {
-    let inbound = InboundMessage::from_qq_group(message);
-    inbound
-        .to_core_request(content)
+    let inbound = platform::qq_official::inbound_from_group(message);
+    platform::to_core_request(&inbound, content)
         .expect("QQ group inbound message should map to CoreRequest")
 }
 
 /// Gateway 侧需要在入队前拿到与 Core 完全一致的 scope_key，用于会话串行调度和 reply cache 隔离。
 pub fn scope_key_from_c2c_message(message: &C2cMessage) -> String {
-    InboundMessage::from_qq_c2c(message)
-        .core_scope_key()
-        .expect("QQ C2C inbound message should have a Core scope")
+    let inbound = platform::qq_official::inbound_from_c2c(message);
+    platform::core_scope_key(&inbound).expect("QQ C2C inbound message should have a Core scope")
 }
 
 /// 群聊 scope 直接复用 Core 的 `group:{group_id}` 规则，避免 Gateway 自己维护第二套会话边界。
 pub fn scope_key_from_group_message(message: &GroupMessage) -> String {
-    InboundMessage::from_qq_group(message)
-        .core_scope_key()
-        .expect("QQ group inbound message should have a Core scope")
+    let inbound = platform::qq_official::inbound_from_group(message);
+    platform::core_scope_key(&inbound).expect("QQ group inbound message should have a Core scope")
 }
 
 /// Egress 层是 gateway 内唯一允许拼接 Core 文本协议的位置。
 /// 这里把 reply block 和附件备注按既有协议收口，避免平台字段污染 Core 稳定模型。
 pub fn build_respond_content(message: &C2cMessage) -> String {
-    InboundMessage::from_qq_c2c(message).render_text_for_core()
+    let inbound = platform::qq_official::inbound_from_c2c(message);
+    platform::render_text_for_core(&inbound)
 }
 
 pub fn build_group_respond_content(message: &GroupMessage, active_keywords: &[String]) -> String {
     let content = normalize_group_command_content(&message.content, active_keywords);
-    let mut inbound = InboundMessage::from_qq_group(message);
+    let mut inbound = platform::qq_official::inbound_from_group(message);
     inbound.text = content;
-    inbound.render_text_for_core()
+    platform::render_text_for_core(&inbound)
 }
 
 fn normalize_group_command_content(content: &str, active_keywords: &[String]) -> String {
